MIARA KIERO KEKE

+++ Scripts/EnemyStats.cs
@@ -12,17 +12,17 @@ namespace ProyectoSDL2.Engine.Scripts
 
         private int hpEnemy;
         private int dmgEnemy;
-        private int moveSpeedEnemy;
+        private float moveSpeedEnemy;
         private int armorEnemy;
 
         //Gets
 
         public int HpEnemy => hpEnemy;
         public int DmgEnemy => dmgEnemy;
-        public int MoveSpeedEnemy => moveSpeedEnemy;
+        public float MoveSpeedEnemy => moveSpeedEnemy;
         public int ArmorEnemy => armorEnemy;
 
-       public EnemyStats(int hpEnemy, int dmgEnemy, int moveSpeedEnemies, int armorEnemies)
+       public EnemyStats(int hpEnemy, int dmgEnemy, float moveSpeedEnemies, int armorEnemies)
         {
             this.hpEnemy = hpEnemy;
             this.dmgEnemy = dmgEnemy;
